chore: manage block conversion logic in blocks store
