+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/tests/Code.test.tsx
@@ -68,4 +68,96 @@ describe('Code', () => {
       },
     ]);
   });
+
+  it('converts a quote block to a code block', () => {
+    const baseEditor = createEditor();
+    baseEditor.children = [
+      {
+        type: 'quote',
+        children: [
+          {
+            type: 'text',
+            text: 'Some quote',
+          },
+        ],
+      },
+    ];
+
+    Transforms.select(baseEditor, {
+      anchor: { path: [0, 0], offset: 0 },
+      focus: { path: [0, 0], offset: 0 },
+    });
+
+    codeBlocks.code.handleConvert(baseEditor);
+
+    expect(baseEditor.children).toEqual([
+      {
+        type: 'code',
+        children: [
+          {
+            type: 'text',
+            text: 'Some quote',
+          },
+        ],
+      },
+      // Should insert a new paragraph as it was the last block
+      {
+        type: 'paragraph',
+        children: [{ type: 'text', text: '' }],
+      },
+    ]);
+  });
+
+  it('should not insert an empty block below if the converted block is not the last one', () => {
+    const baseEditor = createEditor();
+    baseEditor.children = [
+      {
+        type: 'quote',
+        children: [
+          {
+            type: 'text',
+            text: 'Some quote',
+          },
+        ],
+      },
+      {
+        type: 'paragraph',
+        children: [
+          {
+            type: 'text',
+            text: 'Some paragraph',
+          },
+        ],
+      },
+    ];
+
+    Transforms.select(baseEditor, {
+      anchor: { path: [0, 0], offset: 0 },
+      focus: { path: [0, 0], offset: 0 },
+    });
+
+    codeBlocks.code.handleConvert(baseEditor);
+
+    expect(baseEditor.children).toEqual([
+      {
+        type: 'code',
+        children: [
+          {
+            type: 'text',
+            text: 'Some quote',
+          },
+        ],
+      },
+      {
+        type: 'paragraph',
+        children: [
+          {
+            type: 'text',
+            text: 'Some paragraph',
+          },
+        ],
+      },
+      // Nothing should be inserted here
+    ]);
+  });
 });
