+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/tests/Heading.test.tsx
@@ -1,4 +1,6 @@
+/* eslint-disable testing-library/no-node-access */
 import { render, screen } from '@testing-library/react';
+import { Transforms, createEditor } from 'slate';
 
 import { headingBlocks } from '../Heading';
 
@@ -22,5 +24,27 @@ describe('Heading', () => {
 
     const heading = screen.getByRole('heading', { level: 2, name: 'Some heading' });
     expect(heading).toBeInTheDocument();
+  });
+
+  it('converts a code block to a heading', () => {
+    const baseEditor = createEditor();
+    baseEditor.children = [
+      {
+        type: 'code',
+        children: [
+          {
+            type: 'text',
+            text: 'Line of code',
+          },
+        ],
+      },
+    ];
+
+    Transforms.select(baseEditor, {
+      anchor: { path: [0, 0], offset: 0 },
+      focus: { path: [0, 0], offset: 0 },
+    });
+
+    headingBlocks['heading-six'].handleConvert(baseEditor);
   });
 });
