+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/tests/Wrapper.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { lightTheme } from '@strapi/design-system';
 import { IntlProvider } from 'react-intl';
-import { createEditor } from 'slate';
+import { type Editor, createEditor } from 'slate';
 import { Slate, withReact } from 'slate-react';
 import { ThemeProvider } from 'styled-components';
 
@@ -15,13 +15,14 @@ import { listBlocks } from '../List';
 import { paragraphBlocks } from '../Paragraph';
 import { quoteBlocks } from '../Quote';
 
-const baseEditor = createEditor();
+const defaultBaseEditor = createEditor();
 
 interface WrapperProps {
   children: React.ReactNode;
+  baseEditor?: Editor;
 }
 
-const Wrapper = ({ children }: WrapperProps) => {
+const Wrapper = ({ children, baseEditor = defaultBaseEditor }: WrapperProps) => {
   const [editor] = React.useState(() => withReact(baseEditor));
 
   const blocks: BlocksStore = {
@@ -37,7 +38,7 @@ const Wrapper = ({ children }: WrapperProps) => {
   return (
     <ThemeProvider theme={lightTheme}>
       <IntlProvider messages={{}} locale="en">
-        <Slate initialValue={[]} editor={editor}>
+        <Slate initialValue={baseEditor.children} editor={editor}>
           <BlocksEditorProvider blocks={blocks} disabled={false}>
             {children}
           </BlocksEditorProvider>
